[mysql-integration] Replace int type with long type for ids Compare encrypted passwords in login method

+++ tesisCdagAsobiguaApi/Domain/Models/Login.cs
@@ -3,13 +3,13 @@ namespace tesisCdagAsobiguaApi.Domain.Models
 {
     public class Login
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public DateTime TimeStamp { get; set; }
 
-        public int TrainerId { get; set; }
+        public long TrainerId { get; set; }
         public User Trainer { get; set; }
 
-        public int PlayerId { get; set; }
+        public long PlayerId { get; set; }
         public User Player { get; set; }
     }
 }
